:sparkles:  poser base au début + fix camera direction (#108)
* 🚚 Move Widows manager files in folder

* ✨ Implement block building and destruction functionality in BuilderManager

* ✨ Add support for Base block in BuilderManager copy functionality

* ✨ Add cancel functionality to BuilderManager for block building

* ✨ Prevent destruction of Base blocks in BuilderManager

* ✨ Reverse Base block value and add TextBaseGuide texture

* 🔥  Remove TextBaseGuide texture

* ✨ Fix camera movement direction and invert zoom scaling

* ✨ Fix mouse position calculation in BuilderManager

+++ src/Game/Map/BuilderManager.cpp
@@ -1,11 +1,12 @@
-#include "BuilderManager.hpp"
-#include <Tapis.hpp>
 /*
 ** EPITECH PROJECT, 2025
 ** FactoryHub
 ** File description:
 ** BuilderManager
 */
+#include "BuilderManager.hpp"
+#include "Tapis.hpp"
+#include "Base.hpp"
 
 BuilderManager *BuilderManager::instance = nullptr;
 
@@ -39,49 +40,87 @@ static bool isMouseClicked(GLFWwindow *window, int key, int lastKeyState)
 
 void BuilderManager::updateKeyState(GLFWwindow *window, MapGrid &map)
 {
-    if (isKeyClicked(window, GLFW_KEY_B, _lastKeyStates[GLFW_KEY_B]))
+    updateBuildKeys(window, map);
+    updateDestroyKeys(window, map);
+    updateRotateKeys(window, map);
+    updateCancelKeys(window, map);
+    for (int key : _keys)
+        _lastKeyStates[key] = updateLastKeyState(key, window, _lastKeyStates[key]);
+    for (int key : _mouseKeys)
+        _lastKeyStates[key] = updateLastMouseState(key, window, _lastKeyStates[key]);
+    updateSprite(window);
+}
+
+void BuilderManager::updateBuildKeys(GLFWwindow *window, MapGrid &map)
+{
+    if (isKeyClicked(window, GLFW_KEY_B, _lastKeyStates[GLFW_KEY_B])) {
+        if (_isBase == true && _isBuilding == BUILD)
+            return;
         set_isBuilding((_isBuilding == BUILD) ? NONE : BUILD);
+    }
+    if (isMouseClicked(window, GLFW_MOUSE_BUTTON_1, _lastKeyStates[GLFW_MOUSE_BUTTON_1]) && _isBuilding == BUILD){
+        buildBlock(window, map);
+        if (_isBase == true) {
+            set_isBuilding(NONE);
+            _isBase = false;
+            blockBuilding = std::make_shared<Tapis>();
+        }
+    }
+}
+
+void BuilderManager::updateDestroyKeys(GLFWwindow *window, MapGrid &map)
+{
     if (isKeyClicked(window, GLFW_KEY_V, _lastKeyStates[GLFW_KEY_V]))
         set_isBuilding((_isBuilding == DESTROY) ? NONE : DESTROY);
+    if (isMouseClicked(window, GLFW_MOUSE_BUTTON_1, _lastKeyStates[GLFW_MOUSE_BUTTON_1]) && _isBuilding == DESTROY)
+        destroyBlock(window, map);
+}
+
+void BuilderManager::updateRotateKeys(GLFWwindow *window, MapGrid &map)
+{
     if (isKeyClicked(window, GLFW_KEY_R, _lastKeyStates[GLFW_KEY_R]) && _isBuilding == BUILD) {
         _direction = (Direction)((_direction + 3) % 4);
         _spriteBuild->setDirection(_direction * 90);
     }
-    if (isMouseClicked(window, GLFW_MOUSE_BUTTON_2, _lastKeyStates[GLFW_MOUSE_BUTTON_2]) && _isBuilding != NONE)
+}
+
+void BuilderManager::updateCancelKeys(GLFWwindow *window, MapGrid &map)
+{
+    if (isMouseClicked(window, GLFW_MOUSE_BUTTON_2, _lastKeyStates[GLFW_MOUSE_BUTTON_2])
+        && _isBuilding != NONE) {
+        if (_isBase == true && _isBuilding == BUILD)
+            return;
         set_isBuilding(NONE);
-    if (isMouseClicked(window, GLFW_MOUSE_BUTTON_1, _lastKeyStates[GLFW_MOUSE_BUTTON_1]) && _isBuilding == BUILD)
-        buildBlock(window, map);
-    if (isMouseClicked(window, GLFW_MOUSE_BUTTON_1, _lastKeyStates[GLFW_MOUSE_BUTTON_1]) && _isBuilding == DESTROY)
-        destroyBlock(window, map);
-    _lastKeyStates[GLFW_KEY_B] = updateLastKeyState(GLFW_KEY_B,
-        window, _lastKeyStates[GLFW_KEY_B]);
-    _lastKeyStates[GLFW_KEY_V] = updateLastKeyState(GLFW_KEY_V,
-        window, _lastKeyStates[GLFW_KEY_V]);
-    _lastKeyStates[GLFW_KEY_R] = updateLastKeyState(GLFW_KEY_R,
-        window, _lastKeyStates[GLFW_KEY_R]);
-    _lastKeyStates[GLFW_MOUSE_BUTTON_2] = updateLastMouseState(GLFW_MOUSE_BUTTON_2,
-        window, _lastKeyStates[GLFW_MOUSE_BUTTON_2]);
-    _lastKeyStates[GLFW_MOUSE_BUTTON_1] = updateLastMouseState(GLFW_MOUSE_BUTTON_1,
-        window, _lastKeyStates[GLFW_MOUSE_BUTTON_1]);
+    }
+}
+
+void BuilderManager::updateSprite(GLFWwindow *window)
+{
     if (_isBuilding == BUILD)
         _spriteBuild->setPosition(glm::vec3(getMousePos(window), 0.0f));
     if (_isBuilding == DESTROY)
         _spriteDestroy->setPosition(glm::vec3(getMousePos(window), 0.0f));
 }
 
+void BuilderManager::setIsBase(bool isBase)
+{
+    _isBase = isBase;
+}
+
 BuilderManager::BuilderManager()
 {
+    _keys = {GLFW_KEY_B, GLFW_KEY_V, GLFW_KEY_R};
+    _mouseKeys = {GLFW_MOUSE_BUTTON_1, GLFW_MOUSE_BUTTON_2};
     if (instance == nullptr)
         instance = this;
     _spriteBuild = new sdf::Sprite(glm::vec3(0.0f, 0.0f, 0.0f),
         sdf::GetterTextures::instance->getTexture("BuildGhost"), 0.0f);
     _spriteDestroy = new sdf::Sprite(glm::vec3(0.0f, 0.0f, 0.0f),
         sdf::GetterTextures::instance->getTexture("DestroyGhost"), 0.0f);
-    _lastKeyStates[GLFW_KEY_B] = GLFW_RELEASE;
-    _lastKeyStates[GLFW_KEY_V] = GLFW_RELEASE;
-    _lastKeyStates[GLFW_KEY_R] = GLFW_RELEASE;
-    _lastKeyStates[GLFW_MOUSE_BUTTON_2] = GLFW_RELEASE;
-    _lastKeyStates[GLFW_MOUSE_BUTTON_1] = GLFW_RELEASE;
+    for (int key : _keys)
+        _lastKeyStates[key] = GLFW_RELEASE;
+    for (int key : _mouseKeys)
+        _lastKeyStates[key] = GLFW_RELEASE;
     blockBuilding = std::make_shared<Tapis>();
 }
 
@@ -119,6 +158,8 @@ void BuilderManager::destroyBlock(GLFWwindow *window, MapGrid &map)
         blocks = map.getAllBlocksAtPos(pos.x, pos.y);
         if (blocks.size() == 0 || dynamic_cast<ABuilds *>(blocks[blocks.size() - 1].get()) == nullptr)
             return;
+        if (dynamic_cast<Base *>(blocks[blocks.size() - 1].get()) != nullptr)
+            return;
         map.deleteBlock(pos.x, pos.y, blocks.size() - 1);
     } catch (std::exception &e) {}
 }
@@ -142,7 +183,7 @@ glm::vec2 BuilderManager::getMousePos(GLFWwindow *window)
     glm::vec2 vec = sdf::Camera::instance->getPosition();
     xpos = std::round((width / 2) - xpos) / (50 * sdf::Camera::instance->getZoom());
     ypos = std::round((height / 2) - ypos) / (50 * sdf::Camera::instance->getZoom());    
-    pos.x = (int)(-vec.x - (xpos - 0.5f));
+    pos.x = (int)(-vec.x + (xpos + 0.5f));
     pos.y = (int)(-vec.y + (ypos + 0.5f));
     return pos;
 }
@@ -151,10 +192,13 @@ std::shared_ptr<IBlock> BuilderManager::getCopyBlockBuilding()
 {
     if (dynamic_cast<Tapis *>(blockBuilding.get()))
         return std::make_shared<Tapis>();
+    if (dynamic_cast<Base *>(blockBuilding.get()))
+        return std::make_shared<Base>();
     return blockBuilding;
 }
 
-void BuilderManager::setBlockBuilding(std::shared_ptr<IBlock> block)
+void BuilderManager::setBlockBuilding(std::shared_ptr<IBlock> block, bool isBase)
 {
     blockBuilding = block;
+    _isBase = isBase;
 }
